chore: refactor code and files for did:ebsi legal entities
