+++ packages/snap/src/did/ebsi/src/types/ebsiProviderTypes.ts
